Revert "feat(suite-native): add loading while waiting for trezor to show address (#10370)"
This reverts commit 2a55f88a0c769bb7f0293391b653f4c3246d39c1.
