+++ suite-native/qr-code/src/components/QRCode.tsx
@@ -14,7 +14,7 @@ const SCREEN_WIDTH = Dimensions.get('screen').width;
 const MAX_QRCODE_SIZE = 250;
 const QRCODE_PADDING = 12;
 
-export const QRCODE_SIZE =
+const QRCODE_SIZE =
     SCREEN_WIDTH < MAX_QRCODE_SIZE + QRCODE_PADDING ? SCREEN_WIDTH : MAX_QRCODE_SIZE;
 
 const qrCodeContainerStyle = prepareNativeStyle(_ => ({
