+++ suite-native/qr-code/src/index.ts
@@ -1,4 +1,3 @@
 export * from './components/QRCodeScanner';
 export * from './components/AddressQRCode';
 export * from './components/XpubQRCodeBottomSheet';
-export * from './components/QRCode';
