+++ suite-native/receive/src/components/ConfirmOnTrezorImage.tsx
@@ -1,5 +1,5 @@
 import { Image, Pressable } from 'react-native';
-import Animated, { SlideOutDown, SlideInDown } from 'react-native-reanimated';
+import Animated, { FadeIn } from 'react-native-reanimated';
 import React, { useMemo, useState } from 'react';
 
 import { prepareNativeStyle, useNativeStyles } from '@trezor/styles';
@@ -29,11 +29,9 @@ export const ConfirmOnTrezorImage = () => {
 
     return (
         <>
-            <Animated.View
-                entering={SlideInDown}
-                exiting={SlideOutDown}
-                style={applyStyle(imageContainerStyle)}
-            >
+            {/* TODO: images will be revisited in issue:
+            https://github.com/trezor/trezor-suite/issues/9777 */}
+            <Animated.View entering={FadeIn} style={applyStyle(imageContainerStyle)}>
                 <Pressable onPress={handleImagePress}>
                     <Image source={imageSource} />
                 </Pressable>
