+++ suite-native/receive/src/components/DeviceScreenContent.tsx
@@ -12,20 +12,17 @@ import {
 import { prepareNativeStyle, useNativeStyles } from '@trezor/styles';
 import { DeviceModelInternal } from '@trezor/connect';
 import { selectDeviceModel } from '@suite-common/wallet-core';
-import { useTranslate } from '@suite-native/intl';
 
 import { DeviceScreenPagination } from './DeviceScreenPagination';
-import { DEVICE_SCREEN_BACKGROUND_COLOR, DEVICE_TEXT_COLOR } from '../../constants';
-import { parseAddressToDeviceLines } from '../../utils';
-import { DevicePaginationActivePage, isPaginationCompatibleDeviceModel } from '../../types';
-import { ReceiveProgressStep, isAddressRevealed } from '../../hooks/useReceiveProgressSteps';
-import { DeviceScreenLoading } from './DeviceScreenLoading';
+import { DEVICE_SCREEN_BACKGROUND_COLOR, DEVICE_TEXT_COLOR } from '../constants';
+import { parseAddressToDeviceLines } from '../utils';
+import { DevicePaginationActivePage, isPaginationCompatibleDeviceModel } from '../types';
 
 type DeviceScreenContentProps = {
     address: string;
+    isAddressRevealed: boolean;
     isPaginationEnabled: boolean;
     activePage: DevicePaginationActivePage;
-    receiveProgressStep: ReceiveProgressStep;
 };
 
 type DeviceModelLayoutProps = {
@@ -38,21 +35,21 @@ type DeviceModelLayoutProps = {
 
 const deviceToContentStyles = {
     [DeviceModelInternal.T1B1]: {
-        fontSource: require('@trezor/theme/fonts/PixelOperatorMono8-Regular.ttf'),
+        fontSource: require('../../../../packages/theme/fonts/PixelOperatorMono8-Regular.ttf'),
         fontSize: 14,
         lineWidth: 295,
         lineHeight: 17,
         pagerOffset: 0,
     },
     [DeviceModelInternal.T2T1]: {
-        fontSource: require('@trezor/theme/fonts/RobotoMono-Regular.ttf'),
+        fontSource: require('../../../../packages/theme/fonts/RobotoMono-Regular.ttf'),
         fontSize: 20,
         lineWidth: 230,
         lineHeight: 25,
         pagerOffset: 60,
     },
     [DeviceModelInternal.T2B1]: {
-        fontSource: require('@trezor/theme/fonts/PixelOperatorMono8-Regular.ttf'),
+        fontSource: require('../../../../packages/theme/fonts/PixelOperatorMono8-Regular.ttf'),
         fontSize: 14,
         lineWidth: 265,
         lineHeight: 25,
@@ -66,20 +63,20 @@ type ContentCanvasStyleProps = {
     numberOfLines: number;
     isPaginationEnabled: boolean;
     pagerOffset: number;
-    isLoading: boolean;
 };
 
 const contentCanvasStyle = prepareNativeStyle<ContentCanvasStyleProps>(
-    (_, { lineWidth, lineHeight, isPaginationEnabled, numberOfLines, pagerOffset, isLoading }) => ({
+    (_, { lineWidth, lineHeight, isPaginationEnabled, numberOfLines, pagerOffset }) => ({
         width: lineWidth,
         height: lineHeight * numberOfLines,
+
         extend: [
             {
                 condition: isPaginationEnabled,
                 style: {
                     // IF pagination is enabled, the content canvas should be 4 lines high
                     // so the content of the screen does not jump while switching pages.
-                    height: lineHeight * (4 + (isLoading ? 1 : 0)),
+                    height: lineHeight * 4,
                     width: lineWidth + pagerOffset,
                 },
             },
@@ -89,31 +86,21 @@ const contentCanvasStyle = prepareNativeStyle<ContentCanvasStyleProps>(
 
 export const DeviceScreenContent = ({
     address,
+    isAddressRevealed,
     isPaginationEnabled,
     activePage,
-    receiveProgressStep,
 }: DeviceScreenContentProps) => {
     const { applyStyle } = useNativeStyles();
     const deviceModel = useSelector(selectDeviceModel);
-    const { translate } = useTranslate();
-
     const { fontSource, fontSize, lineWidth, lineHeight, pagerOffset } =
         deviceToContentStyles[deviceModel ?? 'T1B1'];
+
     const deviceFont = useFont(fontSource, fontSize);
-
-    const isLoading = receiveProgressStep === ReceiveProgressStep.LoadingOnTrezor;
-
-    const isConfirmOnTrezorReady =
-        receiveProgressStep === ReceiveProgressStep.ShownUncheckedAddress;
-
-    const isAddressShown = isAddressRevealed(receiveProgressStep);
 
     if (!deviceModel) return null;
 
-    const bchPrefixRemovedAddress = address.replace('bitcoincash:', '');
-
     const addressLines = parseAddressToDeviceLines({
-        address: bchPrefixRemovedAddress,
+        address,
         deviceModel,
         isPaginationEnabled,
         activePage,
@@ -122,8 +109,6 @@ export const DeviceScreenContent = ({
     const isPaginationDisplayed =
         isPaginationEnabled && isPaginationCompatibleDeviceModel(deviceModel);
 
-    const lines = addressLines.length + (isLoading ? 1 : 0);
-
     return (
         <Canvas
             style={applyStyle(contentCanvasStyle, {
@@ -131,17 +116,15 @@ export const DeviceScreenContent = ({
                 lineHeight,
                 isPaginationEnabled,
                 pagerOffset,
-                numberOfLines: lines,
-                isLoading,
+                numberOfLines: addressLines.length,
             })}
         >
             <Group>
                 {addressLines.map((line, index) => {
-                    const isTrimmedContent = !isAddressShown && !isConfirmOnTrezorReady;
                     // hide all lines following the one with fade out gradient.
-                    if (isTrimmedContent && index > 1) return null;
+                    if (!isAddressRevealed && index > 1) return null;
 
-                    const isLineFaded = isTrimmedContent && index === 1;
+                    const isLineFaded = !isAddressRevealed && index === 1;
                     const gradientCenterX = lineWidth / 2;
                     const gradientEndY = lineHeight / 0.5;
 
@@ -171,16 +154,9 @@ export const DeviceScreenContent = ({
                     <DeviceScreenPagination
                         deviceModel={deviceModel}
                         activePage={activePage}
-                        isPaginationShown={isAddressShown}
+                        isAddressRevealed={isAddressRevealed}
                     />
                 )}
-                <DeviceScreenLoading
-                    isLoading={isLoading}
-                    font={deviceFont}
-                    xOffset={lineWidth * 0.3}
-                    yOffset={(lines - 1 + 0.75) * lineHeight}
-                    text={translate('moduleReceive.receiveAddressCard.showAddress.loading')}
-                />
             </Group>
         </Canvas>
     );
