+++ suite-native/receive/src/components/DeviceScreenPagination.tsx
@@ -2,12 +2,12 @@ import { Group, ImageSVG, useSVG } from '@shopify/react-native-skia';
 
 import { DeviceModelInternal } from '@trezor/connect';
 
-import { PaginationCompatibleDeviceModel } from '../../types';
+import { PaginationCompatibleDeviceModel } from '../types';
 
 type DeviceScreenPaginationProps = {
     deviceModel: PaginationCompatibleDeviceModel;
     activePage: 1 | 2;
-    isPaginationShown: boolean;
+    isAddressRevealed: boolean;
 };
 
 type PaginationModelMetrics = {
@@ -27,29 +27,29 @@ type PaginationModelMetrics = {
 
 const deviceModelToSvg = {
     [DeviceModelInternal.T2T1]: {
-        paginationPrefixSvg: require('../../../assets/addressPaginationPrefixT2T1.svg'),
-        paginationSuffixSvg: require('../../../assets/addressPaginationSuffixT2T1.svg'),
+        paginationPrefixSvg: require('../../assets/addressPaginationPrefixT2T1.svg'),
+        paginationSuffixSvg: require('../../assets/addressPaginationSuffixT2T1.svg'),
         paginatorPrefixX: 0,
         paginatorPrefixY: 4,
         paginatorSuffixX: 182.5,
         paginatorSuffixY: 82,
 
-        pagerSvg1: require('../../../assets/pager1T2T1.svg'),
-        pagerSvg2: require('../../../assets/pager2T2T1.svg'),
+        pagerSvg1: require('../../assets/pager1T2T1.svg'),
+        pagerSvg2: require('../../assets/pager2T2T1.svg'),
         pagerX: 260,
         pagerY: 30,
         paginationSvgWidth: 40,
         paginationSvgHeight: 17.5,
     },
     [DeviceModelInternal.T2B1]: {
-        paginationPrefixSvg: require('../../../assets/addressPaginationPrefixT2B1.svg'),
-        paginationSuffixSvg: require('../../../assets/addressPaginationSuffixT2B1.svg'),
+        paginationPrefixSvg: require('../../assets/addressPaginationPrefixT2B1.svg'),
+        paginationSuffixSvg: require('../../assets/addressPaginationSuffixT2B1.svg'),
         paginatorPrefixX: 40,
         paginatorPrefixY: 7.5,
         paginatorSuffixX: 210,
         paginatorSuffixY: 82,
-        pagerSvg1: require('../../../assets/pager1T2B1.svg'),
-        pagerSvg2: require('../../../assets/pager2T2B1.svg'),
+        pagerSvg1: require('../../assets/pager1T2B1.svg'),
+        pagerSvg2: require('../../assets/pager2T2B1.svg'),
         pagerX: 277.5,
         pagerY: 0,
         paginationSvgWidth: 12.5,
@@ -60,7 +60,7 @@ const deviceModelToSvg = {
 export const DeviceScreenPagination = ({
     deviceModel,
     activePage,
-    isPaginationShown,
+    isAddressRevealed,
 }: DeviceScreenPaginationProps) => {
     const {
         paginationPrefixSvg,
@@ -88,7 +88,7 @@ export const DeviceScreenPagination = ({
     return (
         <Group>
             <ImageSVG svg={pagerSvg} x={pagerX} y={pagerY} />
-            {isPaginationShown && (
+            {isAddressRevealed && (
                 <ImageSVG
                     svg={paginationSvg}
                     x={paginationX}
